Create gen_save_audio.c
NOTE:
To achieve audio recording in C, we need to rely on platform-specific libraries for capturing audio. Below is a basic example in C using a hypothetical audio recording library. You need to replace "your_audio_library.h" with the appropriate header file for your platform and implement the recordAudio() function according to your library's API.